Add Projects page. Configure HashRouter.
