+++ src/App.tsx
@@ -5,14 +5,15 @@ import Box from '@mui/material/Box';
 import Link from '@mui/material/Link';
 
 import Header from './components/Header';
-import Home from './components/Home';
+import Home from './routes/Home';
 import { Route, Routes } from 'react-router';
+import Projects from './routes/Projects';
 
 function Copyright() {
   return (
-    <Typography variant="body2" color="text.secondary" align="center" my={3}>
+    <Typography variant='body2' color='text.secondary' align='center' my={3}>
       {'© '}
-      <Link color="inherit" href="https://mikeyoung.tech">
+      <Link color='inherit' href='https://mikeyoung.tech'>
         Mike Young
       </Link>{' '}
       {new Date().getFullYear()}
@@ -23,14 +24,15 @@ function Copyright() {
 export default function App() {
   return (
     <>
-      <link rel="preconnect" href="https://fonts.googleapis.com" />
-      <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin='anonymous' />
-      <link href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap" rel="stylesheet"></link>
+      <link rel='preconnect' href='https://fonts.googleapis.com' />
+      <link rel='preconnect' href='https://fonts.gstatic.com' crossOrigin='anonymous' />
+      <link href='https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap' rel='stylesheet'></link>
       <Header />
       <Container disableGutters>
         <Box my={3}>
           <Routes>
-            <Route path="/" Component={Home} />
+            <Route path='/' Component={Home} />
+            <Route path='/projects' Component={Projects} />
           </Routes>
         </Box>
       </Container>
