+++ src/react-app-env.d.ts
@@ -1,2 +1,1 @@
 /// <reference types="react-scripts" />
-// PUBLIC_URL = 'https://mikeyoung.tech/';
